# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /posts:
    x-swagger-router-controller: posts
    get:
      description: returns all the posts
      operationId: posts
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/PostsResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /post:
    x-swagger-router-controller: post
    post:
      description: add a single post
      operationId: addPost
      parameters:
        - in: "body"
          name: "body"
          description: "The new post added by the author"
          required: true
          schema:
            $ref: "#/definitions/Post"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  SuccessResponse:
    required:
      - message
    properties:
      message:
        type: string
  Comment:
    type: object
    required:
      - id
      - name
      - comment
    properties:
      id:
        type: string
      name:
        type: string
      comment:
        type: string
  Post:
    type: object
    required:
      - id
      - title
      - content
      - datetime
      - comments
      - author
    properties:
      id:
        type: string
      title:
        type: string
      content:
        type: string
      datetime:
        type: string
      comments:
        type: array
        items:
          $ref: "#/definitions/Comment"
      author:
        type: string
  PostsResponse:
    required:
      - posts
    properties:
      posts:
        type: array
        items:
          $ref: "#/definitions/Post"